filter+geom: align Op property constraint value_types with setter args
Several migrated Ops expose float/icl64f setters that wrote into
properties registered as Range<int>.  The legacy path stringified
through and truncated; the post-proxy path (prop(name).value = v)
stored a float in typed_value, leaving ConfigFile's adapter cast to
int on save failing.

Per-Op resolution:

  - ThresholdOp: Range<int>{-255..255} / {0..255} → Range<float>.
    Internal storage is float; setters are float.

  - CannyOp: low/high threshold Range<int>{0..2000} → Range<float>.
    m_lowT / m_highT are icl32f; pre-blur stays Range<int>.

  - UnaryArithmeticalOp::setValue(icl64f): Range<int>{-255..512} →
    Range<float>.  Setter casts icl64f→float at the write site to
    match the stored type (only Range<int> and Range<float> adapters
    are registered; Range<double> is not supported).

  - UnaryCompareOp::setValue / setTolerance(icl64f): same treatment.

  - RansacBasedPoseEstimator::setMinPointsForGoodModel(float): narrowed
    setter + ctor param to int.  The value is semantically an integer
    point count; the fit() reader already parsed it as int.

Full test suite: 659/659 pass.

+++ icl/filter/UnaryArithmeticalOp.cpp
@@ -60,7 +60,7 @@ namespace icl::filter {
       m_eOpType(t), m_dValue(val)
   {
     addProperty("op",utils::prop::menuFromCsv(ARITH_MENU), arithName(t));
-    addProperty("value",utils::prop::Range{.min=-255, .max=512}, val);
+    addProperty("value",utils::prop::Range{.min=-255.f, .max=512.f}, static_cast<float>(val));
     registerCallback([this](const Property &p){
       if(p.name == "op")        m_eOpType = parseArith(p.as<std::string>());
       else if(p.name == "value") m_dValue = p.as<icl64f>();
@@ -68,7 +68,9 @@ namespace icl::filter {
   }
 
   void UnaryArithmeticalOp::setOpType(optype t){ prop("op").value = arithName(t); }
-  void UnaryArithmeticalOp::setValue(icl64f v){ prop("value").value = v; }
+  // Property is Range<float> (only int/float adapters registered); setter API is
+  // icl64f historically, so we narrow here to match the stored type.
+  void UnaryArithmeticalOp::setValue(icl64f v){ prop("value").value = static_cast<float>(v); }
 
   REGISTER_CONFIGURABLE_DEFAULT(UnaryArithmeticalOp);
